Move keygen.sh into /usr/sbin and clean up /etc/shibboleth
Move the keygen.sh script into /usr/sbin/shib-keygen and add a man page
for it since users may want to run it after installation.  Remove the
keys generated as part of the build process, since they'll be for the
build host and hence useless in the package.  Also remove an OS X
configuration file.

Modify keygen.sh to cd to /etc/shibboleth before doing anything so that
its file paths will still be correct and comment the --fqdn change.

debian/libapache2-mod-shib2.shibd.init is created as part of the build,
so remove it in the clean target.