Send the native log to syslog
The default native log appender tries to write to native.log in the
Apache log directory, but since that directory is not writable by the
Apache web user (regular Apache logs are handled by the root process),
the logs normally go nowhere.  Change the default to log to syslog so
the logs at least go somewhere.